third_party/WebKit/LayoutTests/shadow-dom/focus-shadowhost-display-none-expected.txt

Check if shadow host with display:none CSS rule for :focus works. crbug.com/482830

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


when shadow host itself is focused, it should match display:none, lose focus then becomes display:block again.
PASS window.getComputedStyle(host).display is "none"
PASS document.activeElement is host
PASS root.activeElement is null
PASS successfullyParsed is true

TEST COMPLETE
PASS window.getComputedStyle(host).display is "block"
PASS document.activeElement is document.body
PASS root.activeElement is null
when shadow host with delegatesFocus=true has focused element inside the shadow, it should also match display:none, then lose focus and become display:block again.
PASS window.getComputedStyle(host).display is "none"
PASS document.activeElement is host
PASS root.activeElement is input
PASS window.getComputedStyle(host).display is "block"
PASS document.activeElement is document.body
PASS root.activeElement is null
